Add info on note synchronization
The info in the guide is not sufficient to understand the setting. Cf. Bug #1393 (I thought I had a bug because I was using Emacs).

I think adding this bit of information will help users make sense of the feature and not worry if the synchronization does not take place instantaneously.